Consolidate API routers
